HW_18/requirements.txt
asgiref==3.6.0
Django==3.2.15
djangorestframework==3.14.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
python-dotenv==0.21.0
pytz==2022.5
sqlparse==0.4.3
Pillow==9.3.0
typing_extensions==4.4.0